requirements.txt
coverage==5.0.3
Django==3.0.4
django-appconf==1.0.3
django-autoslug==1.9.6
django-compressor==2.4
django-debug-toolbar==2.2
django-haystack==3.0b1
django-ipware==2.1.0
django-mdeditor==0.1.17
django-uuslug==1.2.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
jieba==0.42.1
jsonpickle==1.3
lazy-object-proxy==1.4.3
markdown2==2.3.8
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.6
Pillow==7.0.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
python-memcached==1.59
python-slugify==4.0.0
pytz==2019.3
raven==6.10.0
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
WeRoBot==1.10.1
Whoosh==2.7.4
python-logstash==0.4.6
pysolr==3.8.1
